move code for compatibility to older python version into _typing.py
